backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.7
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
pymongo==4.3.3
pandas==2.0.1
scikit-learn==1.2.2
tensorflow==2.12.0
pycaret==3.0.0
celery==5.2.7
